+++ frontends/mit-open/src/pages/HomePage/HomePageCarousel.tsx
